NEW UPDATE 08:10 PM (13th Dec. 2025)
Added Unique Titles to All the Pages. Moved Marker Tool to /pages.

+++ src/pages/Logout/Logout.jsx
@@ -4,14 +4,22 @@ import { useNavigate } from "react-router-dom";
 
 // Local Modules
 import api from "@utils/api";
+import useHead from "@hooks/Head.jsx";
 
 function Logout() {
+  useHead({
+    title: "Logout | G.S.S.S. Mirzewala",
+  });
   const navigate = useNavigate();
   useEffect(() => {
-    api("POST", "auth/logout", true, {}).then(() => {
-      sessionStorage.clear("loggedIn");
+    if (!sessionStorage.getItem("loggedIn")) {
       navigate("/login");
-    });
+    } else {
+      api("POST", "auth/logout", true, {}).then(() => {
+        sessionStorage.clear("loggedIn");
+        navigate("/login");
+      });
+    }
   }, []);
 }
